expose the hyperdrive archive object in the external api
